drm/i915/gvt: Raise a uevent when Dom 0 is ready for Dom U
HV vendors want to know when Dom 0 is ready to start a Dom U
because they want to start Dom U as early as possible. This
feature informs XenGT module as soon as Dom 0 is ready. In our
example, we raise a uevent from XenGT module but the HV vendors
are free to change the module to do any custom action that they
want.

Change-Id: Ibfdaca65002825e14e15527c386db00f59b372e5
Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>
Reviewed-by: He, Min <min.he@intel.com>
Reviewed-on: https://android.intel.com/613374
Reviewed-by: Jiang, Fei <fei.jiang@intel.com>
Reviewed-by: Dong, Eddie <eddie.dong@intel.com>
Tested-by: Dong, Eddie <eddie.dong@intel.com>